+++ subprojects/parseq-tracevis-server/src/main/java/com/linkedin/parseq/HealthCheckHandler.java
@@ -1,0 +1,26 @@
+package com.linkedin.parseq;
+
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import org.eclipse.jetty.server.Request;
+import org.eclipse.jetty.server.handler.AbstractHandler;
+
+
+/**
+ * This class is for health check to check whether the service is running well or not.
+ */
+public class HealthCheckHandler extends AbstractHandler {
+
+  @Override
+  public void handle(String target, Request baseRequest, HttpServletRequest request, HttpServletResponse response)
+      throws IOException, ServletException {
+    if (target.equals("/admin")) {
+      response.setContentType("text/html;charset=utf-8");
+      response.setStatus(HttpServletResponse.SC_OK);
+      baseRequest.setHandled(true);
+      response.getWriter().println("GOOD");
+    }
+  }
+}