+++ subprojects/parseq-tracevis-server/src/main/java/com/linkedin/parseq/TracevisServer.java
@@ -6,14 +6,13 @@ import java.nio.file.Path;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 
+import org.eclipse.jetty.server.Connector;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.handler.ContextHandler;
-import org.eclipse.jetty.server.handler.ContextHandlerCollection;
+import org.eclipse.jetty.server.ServerConnector;
 import org.eclipse.jetty.server.handler.DefaultHandler;
 import org.eclipse.jetty.server.handler.HandlerList;
 import org.eclipse.jetty.server.handler.ResourceHandler;
-import org.eclipse.jetty.util.resource.Resource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -33,6 +32,7 @@ public class TracevisServer {
   private final String _dotLocation;
   final GraphvizEngine _graphvizEngine;
 
+
   public TracevisServer(final String dotLocation, final int port, final Path baseLocation, final Path heapsterLocation,
       final int cacheSize, final long timeoutMs) {
     _dotLocation = dotLocation;
@@ -46,6 +46,7 @@ public class TracevisServer {
         Runtime.getRuntime().availableProcessors(), Constants.DEFAULT_REAPER_DELAY_MS,
         Constants.DEFAULT_PROCESS_QUEUE_SIZE);
   }
+
 
   public void start()
       throws Exception {
@@ -64,8 +65,9 @@ public class TracevisServer {
 
     _graphvizEngine.start();
 
-    Server server = new Server(_port);
+    Server server = new Server();
     server.setAttribute("org.eclipse.jetty.server.Request.maxFormContentSize", -1);
+    server.setConnectors(getConnectors(server));
 
     TracePostHandler tracePostHandler = new TracePostHandler(_staticContentLocation.toString());
 
@@ -85,6 +87,7 @@ public class TracevisServer {
         new JhatHandler(engine),
         tracePostHandler,
         traceHandler,
+        new HealthCheckHandler(),
         heapsterHandler,
         new DefaultHandler()
         });
@@ -100,5 +103,11 @@ public class TracevisServer {
       scheduler.shutdownNow();
       HttpClient.close();
     }
+  }
+
+  protected Connector[] getConnectors(Server server) {
+    ServerConnector connector = new ServerConnector(server);
+    connector.setPort(_port);
+    return new Connector[] { connector };
   }
 }